+++ client/components/SearchBar/GetAllBussiness.js
@@ -1,11 +1,10 @@
 import React, {useEffect} from 'react'
 import {useSelector, useDispatch} from 'react-redux'
-import {getAllBussiness, getCategory} from '../../store/business/getBusiness'
+import {getAllBussiness} from '../../store/business/getBusiness'
 
 const GetAllBussiness = () => {
   const business = useSelector(state => state.business)
   const dispatch = useDispatch()
-
   useEffect(
     () => {
       dispatch(getAllBussiness())
@@ -13,19 +12,8 @@ const GetAllBussiness = () => {
     [getAllBussiness]
   )
 
-  const handleChange = e => {
-    dispatch(getCategory(e.target.value))
-  }
-
   return (
     <div>
-      <select onChange={handleChange}>
-        <option value="restaurant">Restaurant</option>
-        <option value="music">Music</option>
-        <option value="gym">Gym</option>
-        <option value="barbershop">Barbershop</option>
-        <option value="other">Other</option>
-      </select>
       {business.map(query => (
         <div key={query.id}>
           <h4>Name: {query.name}</h4>
