+++ client/components/SearchBar/SearchBusiness.js
@@ -3,6 +3,7 @@ import {useDispatch} from 'react-redux'
 import {searchBusiness} from '../../store/business/getBusiness'
 import useForm from '../Utility/UseForm'
 import history from '../../history'
+import CategoryFilter from './CategoryFilter'
 
 const SearchBusiness = () => {
   const [values, handleChange] = useForm({search: ''})
@@ -27,6 +28,7 @@ const SearchBusiness = () => {
         onKeyDown={keyDown}
         name="search"
       />
+      <CategoryFilter />
     </div>
   )
 }
